Created a VnV namespace and renamed removed the VV prefix from all classes.
Changed how injection points are managed in the core code and in the
runtime configuration file. In this version, the input file only needs
to contain the names of the injection points the user wishes to
instrument. This minimizes the number of injection points that need to
be stored in the static RunTime object.

Added support for runtime scopes in the input file parser. These runtime
scopes allow the user to control which tests should be run. Each test
can now be supplied a list of scopes on which it can run. In the runtime
block, the user provides a list of scopes to run, and only the tests
within those scopes will be loaded. (Note: this is purely an input file
construct. The core injection library knows nothing of the given runtime
scope.)