Converted the xsd schema to json schema and created an example.
This will be used to validate the usrs json files. This is a raw schema
that does not do any checking beyond basic checks for required flags. In
the future, we could use some of the more advanced features of the
schema to verify the file further, prior to runtime,

In particular, a built in schema verification for checking the parameter
mappings between tests and injection points have been implemented
correctly would be useful.