Fix issues associated with the WITH_LOGGING function.
In this branch, the logging functionality was enhanced in such a way
that users of the VnV library can use it as the logger within there
code. We also added the ability to write the logs to the VnV Output
Engine, so they can be kept and displayed as part of the VnV output
format.

This requirement means that the Logging class itself can no longer
be an optional requirement. Basically, the logger itself should always
be compiled into the VnV library.

For each library using the VnV library, logging can be completely
removed by defining WITHOUT_LOGGING in the VnV.h file imported into
the project. In that case, all calls to the Logger, (i.e.,
VnV_Debug(...), etc) are defined out of the final executable using
Macros.

In the case that a linked library was compiled with Logging turned on,
but the user does not want to see logging from that library, the logging
for a particular library can be ignored using the "packageBlackList"
config option in the Logging configuration file.