Fixed a bug in the injection point call
It turns out I could not use the injection point system as I designed
it. This issue was actually pretty difficult to track down, but
basically I was passing variables to the INJECTION_LOOP_BEGIN
function that were no longer in scope during the INJECTION_LOOP_END
function call.

This is the correct behaviour. That is, it is up to the user to ensure
the parameters in a looped injection point are in scope throughout
the entire loop. This bug kind of proves we need to HAMMER home that
point in the documentation.

The same issues likely arrise when defining the Callback using a C++
lambda. In that case, the interface allows the user to capture variables
for use inside the lambda. Those variables are captured as references,
hence they can do out of scope before the end function is called. Again,
this was the intended functionality -- we just need to hammer home that
it is the case.

The alternative is to require the user to specify the parameters
available at the _ITER and _END calls. I had implenented it that way and
found it a bit tedius.

TODO Can we detect if a pointer is still in scope? Probably not in
general.