Added first attempt at Dakota external package.
This attempt is not tested yet -- we will see how it goes. The main
issue this attempt tries to avoid is that Dakota wants to be in the
driver seat

When I first wrote the iterator interface, I was under the impression
that Dakota worked in a set variables -> run code -> process outputs. That
is not true, rather, dakota calls the function evaluation at multiple times
throughout the process. This makes it impossible to factor it out into the
workflow the VnV iterator needs.

One approach to fix this is we can accept function pointers as parameters
to VnV Injection points. This is already supported, but it doesnt flow
as well in a code base as the current implementation (imo).

This dakota interface spawns a thread to manage the dakota processing
of the algorithm. The main thread spawns the dakota thread, then waits for
that thread to set the next set of input parameters. Once the input parameters
are set, the main thread performs the function evaluation. Once the function evaluation
is done, the main thread signals the dakota thread to indicate the function evaluation
is finished. The dakota thread process the results and either quits (when done) or sets
a new set of input parameters. The process repeats until dakota finishes.