Trying to address an issue with MPI communication and Dakota.
The issue is that Dakota requires a MPI_Comm to initialize. But, we
pass a ICommunicator_ptr.

So, we need a user specific way to enable the plugins to pull out the
communicator from the ICommunicator_ptr.

This commit adds a setPackage function that ensures each communicator
knows its own packageName and keyName (e.g. VNV:mpi  or VNV:serial ).

The goal should be to allow the test plugins to extract the communicator
without needing the header file for the specific communicator plugin.

In the end we want to do something like this

if (comm->package=="VNV" && comm->keyName() =="MPI") {
  MPI_Comm comm = comm->get<MPI_Comm>();
}

But, abstract template functions don't exist and we dont have
direct control over what is actually in the "*data" pointer. For instance,
the MPI_Communicator returns a CommKeeper object that contains the MPI_Comm
inside it (that was a fix for some other issue).