Dakota example now working
This commit fixes the issue outlined in the previous commit regarding
Dakota needing access to the MPI_Comm stored in the MPI_Communicator.

To address this, I have added the "raw" function to the ICommunicator_ptr
interface. This function should return a void* pointer to the communication
object a user would expect from that interface.

For example, the MPICommunicator interface returns a void* pointer to the comm
object.

It is left up to the user to determine what type of class that void* can be cast
to. I dont anticipate having more communicators than the serial communicator and the
MPI communicator, so this should be pretty easy. Users can check the communicator
type by calling ICommunicator::getName() which will return a string <package>:<name>
( e.g., VNV:serial, VNV:mpi ).

The dakota test does this

if ( comm->getName() == "VNV:mpi") {
   MPI_Comm* c = (MPI_Comm*) comm->raw();
   ...
} else {
  ...
}

The downside of this is that the tests become specific to a particular set of communicators.
As far as I see it, that is unavoidable. E.g. dakota supports serial and MPI communication. Some
other communication object can not be used with dakota, so we would need to check that anyway.

This commit also updates the dakota input string to accept several parameters, including the
comm size. Setting processors_per_analaysis to commSize ensures dakota does not try to do algorithmic
parallelism

TODO:
   1.  add options for setting the tolerances,methods,etc through the input file
   2.  add other input file options for different parameter searches
   3.  pipe dakota outputs to the outputEngine and write some report content comments
   4.  add support for specifying the gradients and hessians if analytic hessians are
       available

Notes:
   1. Dakota has a Data api for setting the input parameters programatically. The API seems
      pretty nice, but from a simplicity perspective, setting the parameters using the input
      file string is very easy and very descriptive. I doubt the Data API is noticibly faster
      when compared to the execution time. We could look at changing to the data api in the future, especially
      as we add more options. (see dataka-src-> libray_mode.cpp -> dakota_parse_data )