A start on a simple plugin manager for the toolkit.
This script provides a number of options for maintaining a .vnv-plugins file
in the $HOME directory. This is a json file containing the name, package and
version of each VnV plugin the user adds. The file also contains a "revert" history
that allows users to undo changes made to the plugins file. A "default" property
is also available to allow switching between the default "version" of each plugin.

The toolkit will automatically search this directory for any plugins during
loading. The toolkit will load the "default" plugin. we