Macros that make dealing with template parameters easier. This is a stop gap measure until I can come up with something simpler.
The idea with these macros is that for template parameters in tests, the developer should
only need to write the body of the test once. These macros take a body, and then repeat it a bunch
of times for each possible parameter type given.

To test, open templates.h and set line one to ->   \#define TESTING 1
Then, from the terminal, type "cpp templates.h"