Adds an application description property
This commit adds an application description to the input file interface.
The application description is extracted from the comment above the
applications INJECTION_EXECUTABLE macro

When  a user loads a valid vnv executable (as determined by our ability
to extract a schema from it using the provided command) the interface
will now automatically populate the application description  and the
application specification.

The executable path should turn green when the application is valid.

Also, this commit fixes the broken autocomplete support on the
application path field. Application path should now autocomplete
as the user types. For a remote connection (assuming its connected) auto
complete should autocomplete based on the remote file system.