Python Content in Plotly body
The content of the vnv-plotly direcitve (body) is now evaluated using
eval during runtime. Once evaluated, the body must be a python
dictionary mapping strings to valid Json.

This can be used to generate data as needed for plots. For instance, the
Euler example introduction uses this functionality to plot the exact
solution without running the application.

The plotly directive calls options.update(evaled_dict) prior to
rendering it, so the keys of the dict should be valid plotly keys. 

TODO: We currently ignore invalid python content -- We really should
tell the user (or at least the developer who wrote it).

The eval method makes numpy (as np), math (as math) and json (as json)
available to the user for their computations. 